[update] v1.1 spring security 接入数据库,支持权限校验

+++ gengzi_spring_security/src/main/java/fun/gengzi/gengzi_spring_security/sys/dao/SysUsersDao.java
@@ -2,12 +2,17 @@ package fun.gengzi.gengzi_spring_security.sys.dao;
 
 import fun.gengzi.gengzi_spring_security.sys.entity.SysUsers;
 import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.data.jpa.repository.Modifying;
+import org.springframework.data.jpa.repository.Query;
 import org.springframework.stereotype.Repository;
-
+import org.springframework.transaction.annotation.Transactional;
 
 @Repository
-public interface SysUsersDao extends JpaRepository<SysUsers, Integer> {
+public interface SysUsersDao extends JpaRepository<SysUsers, Long> {
 
     SysUsers findByUsername(String username);
+
+
+
 
 }
